+++ ChangeLog.md
@@ -19,6 +19,10 @@
 - Added `set_pen`. ([#7](https://github.com/davep/textual-canvas/pull/7))
 - Fixed off-by-one issue with `draw_rectangle`.
   ([#10](https://github.com/davep/textual-canvas/pull/10))
+- Added `Canvas.clear_pixels`.
+  ([#11](https://github.com/davep/textual-canvas/pull/11))
+- Added `Canvas.clear_pixel`.
+  ([#11](https://github.com/davep/textual-canvas/pull/11))
 
 ## v0.2.0